wdqs-gateway, namespace routing (remove rewriting)
Rewriting has even more problems than first identified,
due to things like https://w.wiki/AMq

This query assumes that the user knows the length of the URI.
However the previous plan for rewriting would have broken this..

I decided in Banff to just got for the namepace for the MVP,
and if a plan comes about for internal namespaces, then that shouldnt
be too hard to make work with real URIs also in the blazegraph
backends...

Most of the tests in the gateway will be broken after this commit and I
should probably just delete them...

ALSO this change makes the query service UI work for localhost testing
and makes a new releaseof it (not really needed in prod).

Now to go back to the API and stop generating internal hosts for
qs namespaces? :(